authorization implemented

+++ src/Controller/UsersController.php
@@ -25,16 +25,20 @@ class UsersController extends AppController
      */
     public function index()
     {
-        $this->paginate = [
-            'conditions' => [
-                'Users.auth' => 0
-            ]
-        ];
-        $users = $this->paginate($this->Users);
-        $title = 'User|List';
+        if($this->request->session()->read('currentUser')['auth'] > 0) {
+            $this->paginate = [
+                'conditions' => [
+                    'Users.auth' => 0
+                ]
+            ];
+            $users = $this->paginate($this->Users);
+            $title = 'User|List';
 
-        $this->set(compact('users', 'title'));
-        $this->set('_serialize', ['users']);
+            $this->set(compact('users', 'title'));
+            $this->set('_serialize', ['users']);
+        } else {
+            return $this->redirect(['controller'=>'Apartments', 'action' => 'index']);
+        }
     }
 
     /**
@@ -44,33 +48,20 @@ class UsersController extends AppController
      */
     public function opindex()
     {
-        $this->paginate = [
+        if($this->request->session()->read('currentUser')['auth'] > 0) {
+            $this->paginate = [
             'conditions' => [
                 'Users.auth' => 1
-            ]
-        ];
-        $users = $this->paginate($this->Users);
-        $title = 'User|List|operators';
+                ]
+            ];
+            $users = $this->paginate($this->Users);
+            $title = 'User|List|operators';
 
-        $this->set(compact('users', 'title'));
-        $this->set('_serialize', ['users']);
-    }
-
-    /**
-     * View method
-     *
-     * @param string|null $id User id.
-     * @return \Cake\Network\Response|null
-     * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
-     */
-    public function view($id = null)
-    {
-        $user = $this->Users->get($id, [
-            'contain' => []
-        ]);
-
-        $this->set('user', $user);
-        $this->set('_serialize', ['user']);
+            $this->set(compact('users', 'title'));
+            $this->set('_serialize', ['users']);
+        } else {
+            return $this->redirect(['controller'=>'Apartments', 'action' => 'index']);
+        }
     }
 
     /**
@@ -80,44 +71,40 @@ class UsersController extends AppController
      */
     public function add()
     {
-        $user = $this->Users->newEntity();
-        $auth = User::AUTH;
-        $title = 'User|Add';
-        if ($this->request->is('post')) {
-            $user = $this->Users->patchEntity($user, $this->request->getData());
-            $mailPassword = $this->request->data('password');
-            $user->password = md5($mailPassword);
-            if ($this->Users->save($user)) {
-                $this->Flash->success(__('The user has been saved.'));
-                if ($user->auth == 0){
-                    $email = new Email();
-                    $email
-                        ->setTemplate('notification')
-                        ->setViewVars(['name' => $user->name, 'validity' => $user->validity, 'username' => $user->username, 'password' => $mailPassword])
-                        ->setLayout('notification')
-                        ->setEmailFormat('html')
-                        ->setSubject('Welcome to Asahi Service Company!')
-                        ->setTo($user->email)
-                        ->setFrom(['admin@asahiservices.com' => 'Asahi Service Company'])
-                        ->send();
-                    return $this->redirect(['action' => 'index']);
-                } else {
-                    return $this->redirect(['action' => 'opindex']);
+        if($this->request->session()->read('currentUser')['auth'] > 1) {
+            $user = $this->Users->newEntity();
+            $auth = User::AUTH;
+            $title = 'User|Add';
+            if ($this->request->is('post')) {
+                $user = $this->Users->patchEntity($user, $this->request->getData());
+                $mailPassword = $this->request->data('password');
+                $user->password = md5($mailPassword);
+                if ($this->Users->save($user)) {
+                    $this->Flash->success(__('The user has been saved.'));
+                    if ($user->auth == 0) {
+                        $email = new Email();
+                        $email
+                            ->setTemplate('notification')
+                            ->setViewVars(['name' => $user->name, 'validity' => $user->validity, 'username' => $user->username, 'password' => $mailPassword])
+                            ->setLayout('notification')
+                            ->setEmailFormat('html')
+                            ->setSubject('Welcome to Asahi Service Company!')
+                            ->setTo($user->email)
+                            ->setFrom(['admin@asahiservices.com' => 'Asahi Service Company'])
+                            ->send();
+                        return $this->redirect(['action' => 'index']);
+                    } else {
+                        return $this->redirect(['action' => 'opindex']);
+                    }
                 }
+                $this->Flash->error(__('The user could not be saved. Please, try again.'));
             }
-            $this->Flash->error(__('The user could not be saved. Please, try again.'));
+            $this->set(compact('user', 'auth', 'title'));
+            $this->set('_serialize', ['user']);
+        } else {
+            return $this->redirect(['controller'=>'Apartments', 'action' => 'index']);
         }
-        $this->set(compact('user', 'auth', 'title'));
-        $this->set('_serialize', ['user']);
     }
-
-//    public function notification(){
-//        $name = 'name';
-//        $validity = '123';
-//        $username = 'username';
-//        $password = 'password';
-//        $this->set(compact('name', 'validity', 'username', 'password'));
-//    }
 
     /**
      * Delete method
@@ -128,13 +115,17 @@ class UsersController extends AppController
      */
     public function delete($id = null)
     {
-        $user = $this->Users->get($id);
-        if ($this->Users->delete($user)) {
-            $this->Flash->success(__('The user has been deleted.'));
-        } else {
-            $this->Flash->error(__('The user could not be deleted. Please, try again.'));
-        }
+        if($this->request->session()->read('currentUser')['auth'] > 0) {
 
+            $user = $this->Users->get($id);
+            if ($user['auth'] > 1){
+                if ($this->Users->delete($user)) {
+                    $this->Flash->success(__('The user has been deleted.'));
+                } else {
+                    $this->Flash->error(__('The user could not be deleted. Please, try again.'));
+                }
+            }
+        }
         return $this->redirect(['action' => 'index']);
     }
 
@@ -150,8 +141,10 @@ class UsersController extends AppController
             $user = $this->Auth->identify();
             $this->request->session()->write('currentUser',$user);
             $this->Auth->setUser($user);
+            if ($user['auth'] == 0){
+                return $this->redirect(['controller' => 'Apartments', 'action' => 'index']);
+            }
             return $this->redirect($this->Auth->redirectUrl());
-            //$this->Flash->error('Login Failed, Please check Username and Password');
         }
     }
